// akadon-frontend/src/app/models/student.model.ts
export class Student {
    studentId?: number;
    email?: string;
    password?: string;
    fullname?: string;
    gender?: Boolean;
    city?: string;
    district?: string;
    ward?: string;
    phoneNumber?: string;
    birthDay?: Date;
    role?: string;
    status?: Boolean;
    image?: string;
    learntCourseNumber?: number;
    lastLogin?: Date;
    rating?: number;
    behaviorPoint?: number;
    verificationCode?: string;
    otpRequestTime?: Date;
}